rosy: Patch libmmcamera2_iface_modules
On camera HAL3 the CDS requests are transferred via functions of the
camera.msm8953.so library to the functions within libmmcamera2_iface_modules.
Since it's hard to shim the right functions of the camera.msm8953.so blob,
patching the libmmcamera2_iface_modules library is easier to suppress
CDS setting.

This is done by patching the relevant value assignment where not
the value is taken of the parameter, but a static value 0x0, which
means the CDS mode is set to OFF.

Change-Id: I391d2d0b94a1356968aaa5c32f2d0255e89c69df